Refactored ResendAuthEmail, deleted dialogs/SaveIdeaWithoutLogin.vue and profilePage/SaveIdeaBookmark.vue as they are abandoned
